+++ yepshop-types/src/main/java/com/yufu/yepshop/types/command/CreateGoodsCommand.java
@@ -7,7 +7,6 @@ import lombok.Getter;
 import lombok.Setter;
 
 import javax.validation.constraints.NotNull;
-import java.math.BigDecimal;
 import java.util.List;
 
 /**
@@ -28,11 +27,11 @@ public class CreateGoodsCommand {
     private String[] urls;
 
     @NotNull(message = "金额不能为空")
-    private BigDecimal price;
+    private Integer price;
 
-    private BigDecimal originalPrice;
+    private Integer originalPrice;
 
-    private BigDecimal postFee;
+    private Integer postFee;
 
     private String categoryId;
